requirements.txt
appnope @ file:///home/conda/feedstock_root/build_artifacts/appnope_1649077682618/work
asgiref==3.6.0
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1670263926556/work
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1618230623929/work
backports.zoneinfo==0.2.1
certifi==2023.5.7
charset-normalizer==3.1.0
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1679481329611/work
contourpy==1.0.7
cycler==0.11.0
debugpy @ file:///Users/runner/miniforge3/conda-bld/debugpy_1680755597432/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
Django==4.2.1
djangorestframework==3.14.0
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1667317341051/work
filelock==3.12.0
fonttools==4.39.4
idna==3.4
importlib-metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1682176699712/work
importlib-resources==5.12.0
ipykernel @ file:///Users/runner/miniforge3/conda-bld/ipykernel_1684162997309/work
ipython @ file:///Users/runner/miniforge3/conda-bld/ipython_1653755011383/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1669134318875/work
Jinja2==3.1.2
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1681432441054/work
jupyter_core @ file:///Users/runner/miniforge3/conda-bld/jupyter_core_1678994269065/work
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1660814786464/work
mpmath==1.3.0
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1664684991461/work
networkx==3.1
numpy==1.24.3
opencv-python==4.7.0.72
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1681337016113/work
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1667297516076/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==9.5.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1683850015520/work
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1677600924538/work
psutil @ file:///Users/runner/miniforge3/conda-bld/psutil_1681775314478/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1681904169130/work
pyparsing==3.0.9
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytz==2023.3
pyzmq @ file:///Users/runner/miniforge3/conda-bld/pyzmq_1679317074020/work
requests==2.31.0
scipy==1.10.1
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sqlparse==0.4.4
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
sympy==1.12
torch==2.0.1
torchvision==0.15.2
tornado @ file:///Users/runner/miniforge3/conda-bld/tornado_1684150353057/work
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1675110562325/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1678559861143/work
urllib3==2.0.2
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1673864653149/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1677313463193/work
